Move ChecksummedBytes to the object module
