+++ mountpoint-s3/src/data_cache.rs
@@ -10,10 +10,10 @@ mod in_memory_data_cache;
 
 use thiserror::Error;
 
-pub use crate::checksums::ChecksummedBytes;
 pub use crate::data_cache::cache_directory::ManagedCacheDir;
 pub use crate::data_cache::disk_data_cache::{CacheLimit, DiskDataCache, DiskDataCacheConfig};
 pub use crate::data_cache::in_memory_data_cache::InMemoryDataCache;
+pub use crate::object::ChecksummedBytes;
 
 use crate::object::ObjectId;
