+++ mountpoint-s3/src/prefetch/part_stream.rs
@@ -6,8 +6,7 @@ use futures::{pin_mut, task::Spawn, StreamExt};
 use mountpoint_s3_client::{types::ETag, ObjectClient};
 use tracing::{debug_span, error, trace, Instrument};
 
-use crate::checksums::ChecksummedBytes;
-use crate::object::{ObjectId, ObjectPart};
+use crate::object::{ChecksummedBytes, ObjectId, ObjectPart};
 use crate::prefetch::part_queue::unbounded_part_queue;
 use crate::prefetch::task::RequestTask;
 use crate::prefetch::PrefetchReadError;
